refactor: extract project data pipeline into shared module
